one more compatible fix
