+++ src/getModule.js
@@ -1,5 +1,5 @@
 const Module = typeof __webpack_require__ === 'function'
-  ? require('../webpack/module')
+  ? require('../webpack/src/module')
   : require('module');
 
 export default Module;
